use a filter to automatically extract path info args from req URI
Assumes that arguments are always integers, but we could
potentially make that more flexible.

This is a nice approach because all servlets and filters need
to do is get the "args" parameter from the request attributes,
and it will have the arguments as a list of integers.
A static method (ServletUtil.getRequestArgs) is provided to
do this.

Also, refactored error responses to support other status codes.