don't attempt to use clojure and makefile brushes
Basically, syntaxhighlighter v4's support for adding brushes
is completely broken.

It claims to support v3 brushes (when building with --compat),
but they don't work.

Adding a v4 brush such as https://github.com/p--q/brush-Makefile
causes the build to fail.

The instructions on the syntaxhighlighter wiki are useless.

I might keep working on this, but given the utter lack of
useful documentation, I probably won't.